Create README.md
This project is the ASP.NET MVC 5 version of an ECommerce site aimed to sell sports features. This is a simple project written with C# langage, ASP.NET MVC 5, .NET 5.0, Entity Framework, 
Bootstrap 4.6. The goal is set up a fictitious Ecommerce web site to sell sports wears and drinks for athletes.